Change create_region_input.py script for topo.nc file to use doubles instead of int.
Turns out the data in the source files is floating point, so
we might as well keep the info!

+++ scripts/create_region_input.py
@@ -192,9 +192,9 @@ def create_template_topo_file(fname, sizey=10, sizex=10):
 
   Y = ncfile.createDimension('Y', sizey)
   X = ncfile.createDimension('X', sizex)
-  slope = ncfile.createVariable('slope', np.int, ('Y', 'X',))
-  aspect = ncfile.createVariable('aspect', np.int, ('Y', 'X',))
-  elevation = ncfile.createVariable('elevation', np.int, ('Y', 'X',))
+  slope = ncfile.createVariable('slope', np.double, ('Y', 'X',))
+  aspect = ncfile.createVariable('aspect', np.double, ('Y', 'X',))
+  elevation = ncfile.createVariable('elevation', np.double, ('Y', 'X',))
 
   ncfile.source = source_attr_string()
   ncfile.close()
